Update document.R
Managed to functionalize the file I/O, etc. to make it able to handle N data-sets, given that it doesn't cause overflow issues within R.
By default it will only make two, but it reads in every set of data in case we want to set the default later.

Default is set to 2 because of aforementioned issues with R allowing you to input any N data-sets, potentially causing memory issues.