I updated the Daemon Code.
I fixed some bugs that were related to starting up the daemon
and creating the PID file.
Download this version instead of the last one.
Now the Daemon fully works and you can start it up and shut it down
and it runs in the background without any problems.
Sanitzing the environmental variables is not currently implemented though.