Update line scanning for input files
There was a sudden change to input file format that required a switch to regex for input file scanning.

This new format meant there wouldn't be standardized positions of key words, so a lot of methods quickly became unviable and the only realistic option left was using regex. While this unfortunately results in adding the library 'stringr' to the dependency list, we're now able to use regex in all other parts of the R-code if needed.