I fully integrated the daemon with the UI 100%.
Now the UI is more tightly together with the daemon.
I replaced all cout statements with syslog() statements.
I also modified the UI to work on any computer, by getting the $HOME directory.
I also created a new version of the Makefile and I deleted the old version of the Makefile.

The GUI code can now run and kill the daemon code successfully.