I fixed a minor bug.
Previously if a parent directory existed, but the child directory did not exist,
the child directory would not have been created, and so the application
would not be able to open that directory to write to it.

I changed it so that even if the parent directory already exists, the child
directory will still be created if it does not already exist.
After that the appliation can use that directory without problems.