I made a big change to the skeleton Makefile.
It turned out that many of the lines of code in the previous version of the Makefile
were no longer needed because they were generated by qmake only, for qmake only.
I got rid of those extraneous lines.

Now the only four lines with the FIXME that we have to replace are:
 - INCPATH
 - LIBS
 - Rule to make the metacode file moc_mainwindow.cpp
 - Rule to make the metacode file ui_mainwindow.h

I also updated .gitignore to make it ignore the generated metacode files.