Implementing passing configured detection settings to the camera code.
I worked on the feature of enabling the user to toggle various detection
settings in the GUI, and passing those parameters to the camera code.
I successfully implemented the backend infrastructure to handle this feature.
Now I am waiting for Slavik to implement these settings in the camera code.
Also the GUI has a bug that check boxes are unresponsive to clicks.
The internal states of the check boxes are not being updated.
I am waiting for Wei Tao to fix that bug.